Better directory structure for autogradpp
