+++ include/detail.h
@@ -1,0 +1,58 @@
+#pragma once
+
+#include <map>
+#include <memory>
+
+#include "torch/csrc/autograd/engine.h"
+#include "torch/csrc/autograd/grad_mode.h"
+
+#define AUTOGRAD_OPTIMIZER_CLASS(Type) class Type : public autograd::Optimizer_CRTP<Type>
+#define AUTOGRAD_KWARG(CLS, TYP, NAME, DEFAULT, OPTION) \
+  TYP NAME ## _ = DEFAULT; \
+  CLS & NAME(TYP x = OPTION) { NAME ## _ = x; return *this; }
+
+
+namespace {
+namespace tag = torch::autograd;
+using IntVec = decltype(std::declval<at::IntList>().vec());
+}
+
+namespace autograd {
+namespace detail {
+extern tag::Engine engine;
+}
+
+class ContainerImpl;
+class OptimizerImpl;
+using Variable = tag::Variable;
+using variable_list = tag::variable_list;
+using Tensor = tag::Tensor;
+using Container = std::shared_ptr<ContainerImpl>;
+using Optimizer = std::shared_ptr<OptimizerImpl>;
+
+void backward(Tensor loss, bool keep_graph=false);
+
+inline Variable Var(at::Tensor data, bool requires_grad=true) {
+  return tag::make_variable(data, requires_grad);
+}
+
+// This is thread local!!!
+inline void set_grad_enabled(bool val=true) {
+  tag::GradMode::set_enabled(val);
+}
+
+// RAII thread local lock that stops future execution from building gradients
+class no_grad_guard {
+ public:
+  no_grad_guard() {
+    tag::GradMode::set_enabled(false);
+  }
+
+  ~no_grad_guard() {
+    tag::GradMode::set_enabled(true);
+  }
+};
+
+void setSeed(uint64_t seed);
+
+} // namespace autograd
