+++ tests/mnist/t10k-labels-idx1-ubyte
